RK user stat processing in jupyter notebook
This is the full raw code with all the blemishes and warts to go along.  You can see my thought process and that I pull in raw logs from keycloak, as well as jira tickets and a raw json from rk to build out the user lists and map to the users in the logs to assign orgs to.